Move assertValidProps into shared/utils since it will be used by the server renderer
